+++ @es-labs/node/auth/redis.js
@@ -2,6 +2,6 @@
 
 let redis
 exports.setTokenService = (service) => redis = service
-exports.setRefreshToken = async (id, refresh_token) => await redis.set(id, refresh_token)
-exports.getRefreshToken = async (id) => await redis.get(id)
-exports.revokeRefreshToken = async(id) => await redis.del(id)
+exports.setRefreshToken = async (id, refresh_token) => redis.set(id, refresh_token)
+exports.getRefreshToken = async (id) => redis.get(id)
+exports.revokeRefreshToken = async(id) => redis.del(id)