+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Version 0.6.8
 - [chore] update packages & cleanup & work on improving documentation
 - [js-web/vue-vite] integrate [https://github.com/ais-one/favv/vitevue](https://github.com/ais-one/favv/vitevue) to [js-web/vue-vite](js-web/vue-vite)
+- [js-web/vue-vite] update firebase messaging from 8 to 9, fix PN bugs and improve PN test on UI
+- [js-node/expressjs] update firebase messaging from 8 to 9, fix PN bugs
 - IN PROGRESS
   - [chore] add sri and crossorigin to scripts - https://github.com/bigskysoftware/htmx/issues/261#issuecomment-753850081
   - [chore] migration to node 16 and npm 7